// fashion_web_admin_ui/src/app/api/admin-api.service.generated.ts
export const ADMIN_API_BASE_URL = new InjectionToken<string>('ADMIN_API_BASE_URL');
// ...
export class ProductResponseDTO implements IProductResponseDTO {
    id?: string;
    name?: string | undefined;
    description?: string | undefined;
    price?: number;
    thumbnailUrl?: string | undefined;
    categoryId?: string;

    constructor(data?: IProductResponseDTO) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (this as any)[property] = (data as any)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.id = _data["id"];
            this.name = _data["name"];
            this.description = _data["description"];
            this.price = _data["price"];
            this.thumbnailUrl = _data["thumbnailUrl"];
            this.categoryId = _data["categoryId"];
        }
    }

    static fromJS(data: any): ProductResponseDTO {
        data = typeof data === 'object' ? data : {};
        let result = new ProductResponseDTO();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["id"] = this.id;
        data["name"] = this.name;
        data["description"] = this.description;
        data["price"] = this.price;
        data["thumbnailUrl"] = this.thumbnailUrl;
        data["categoryId"] = this.categoryId;
        return data;
    }
}

export interface IProductResponseDTO {
    id?: string;
    name?: string | undefined;
    description?: string | undefined;
    price?: number;
    thumbnailUrl?: string | undefined;
    categoryId?: string;
}
// ...
export class VariantDTO implements IVariantDTO {
    id?: string;
    productName?: string | undefined;
    size?: string | undefined;
    color?: string | undefined;
    thumbnailUrl?: string | undefined;
    price?: number;
    stockQuantity?: number;

    constructor(data?: IVariantDTO) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (this as any)[property] = (data as any)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.id = _data["id"];
            this.productName = _data["productName"];
            this.size = _data["size"];
            this.color = _data["color"];
            this.thumbnailUrl = _data["thumbnailUrl"];
            this.price = _data["price"];
            this.stockQuantity = _data["stockQuantity"];
        }
    }

    static fromJS(data: any): VariantDTO {
        data = typeof data === 'object' ? data : {};
        let result = new VariantDTO();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["id"] = this.id;
        data["productName"] = this.productName;
        data["size"] = this.size;
        data["color"] = this.color;
        data["thumbnailUrl"] = this.thumbnailUrl;
        data["price"] = this.price;
        data["stockQuantity"] = this.stockQuantity;
        return data;
    }
}

export interface IVariantDTO {
    id?: string;
    productName?: string | undefined;
    size?: string | undefined;
    color?: string | undefined;
    thumbnailUrl?: string | undefined;
    price?: number;
    stockQuantity?: number;
}

export class VariantDTOApiResponse implements IVariantDTOApiResponse {
    success?: boolean;
    message?: string | undefined;
    statusCode?: number | undefined;
    data?: VariantDTO;

    constructor(data?: IVariantDTOApiResponse) {
        if (data) {
            for (var property in data) {
                if (data.hasOwnProperty(property))
                    (this as any)[property] = (data as any)[property];
            }
        }
    }

    init(_data?: any) {
        if (_data) {
            this.success = _data["success"];
            this.message = _data["message"];
            this.statusCode = _data["statusCode"];
            this.data = _data["data"] ? VariantDTO.fromJS(_data["data"]) : undefined as any;
        }
    }

    static fromJS(data: any): VariantDTOApiResponse {
        data = typeof data === 'object' ? data : {};
        let result = new VariantDTOApiResponse();
        result.init(data);
        return result;
    }

    toJSON(data?: any) {
        data = typeof data === 'object' ? data : {};
        data["success"] = this.success;
        data["message"] = this.message;
        data["statusCode"] = this.statusCode;
        data["data"] = this.data ? this.data.toJSON() : undefined as any;
        return data;
    }
}

export interface IVariantDTOApiResponse {
    success?: boolean;
    message?: string | undefined;
    statusCode?: number | undefined;
    data?: VariantDTO;
}
// ...
export interface FileParameter {
    data: any;
    fileName: string;
}